Intercept Groovy properties
Property reads are simple as they go through the same CallSiteArray, we
just add another interceptor that redirects the read to the new
property.

Writes are completely different beasts because there is no CallSite.
Instead lookup goes through metaclass without the middleman. So we have
to intercept at metaclass level as well. Thankfully, a metaclass with
a carefully crafted name can be picked up by Groovy runtime
automagically[1].

We might consider switching to metaclasses for other things to make
things uniform.

For Gradle these metaclasses will have to be generated for classes that
we want to backport at some point.

  [1]: https://docs.groovy-lang.org/latest/html/documentation/core-metaprogramming.html#_magic_package